Resolved duplicate ID issue
Duplicating the two sections used for the services content areas resulted in duplicate IDs.

Resolved this by duplicating the CSS for each ID and adding v2 to the end of the ID in the HTML and CSS.

I think this provided the path of least resistance, and also allowed for changes to be made per section in the future.